Solve 6th flex exercise
Hell yeah! This one was much closer to the final solution.

I managed to get close with little differences in the CSS, such as the
gap between elements, and an error on the border from the two middle
buttons.

I feel like flex finally clicked on this one, and I was able to
employ it in an effective way